Hash aplicavel para qualquer arquivo, parsing do documento feito.

+++ SHA3.py
@@ -222,7 +222,7 @@ def toBits(s):
 
 def fromBits(bits):
     chars = []
-    for b in range(len(bits) / 8):
+    for b in range(int(len(bits) / 8)):
         byte = bits[b*8:(b+1)*8]
         chars.append(chr(int(''.join([str(bit) for bit in byte]), 2)))
     return ''.join(chars)
